Move contents of `nopJitter` and `randomJitter` to the `jitter` file
