--- v2/nop_jitter.go
@@ -1,11 +1,0 @@
-package backoff
-
-type nopJitter struct{}
-
-func newNopJitter() *nopJitter {
-	return &nopJitter{}
-}
-
-func (j *nopJitter) apply(interval float64) float64 {
-	return interval
-}
